+++ src/common/index.ts
@@ -1,6 +1,6 @@
 export * from "./configuration";
 export * from "./xqlint";
-export * from "./create-document-selector";
+export * from "./document-selectors";
 export * from "./extension-state";
 export * from "./native-commands";
 export * from "./xml-traverser";